base: cvd: Cast packed struct header_version to avoid reference binding
Passing a member of a packed struct directly into fmt::format /
CF_ERRF fails to compile because a packed field cannot be bound to a
const reference parameter.
Cast header_version to uint32_t to avoid reference binding.

+++ base/cvd/cuttlefish/host/commands/assemble_cvd/boot_image/boot_image.cc
@@ -61,7 +61,8 @@ Result<BootImage> BootImage::Read(std::unique_ptr<ReaderSeeker> rd) {
     case 4:
       return BootImage(std::move(rd), *reinterpret_cast<boot_img_hdr_v4*>(&v2));
     default:
-      return CF_ERRF("Unknown header version '{}'", v2.header_version);
+      return CF_ERRF("Unknown header version '{}'",
+                     static_cast<uint32_t>(v2.header_version));
   }
 }
